langchain/requirements.txt
langchain
PyPDF2
python-dotenv
requests
chromadb
termcolor
